Fix MSVC warnings in utils\*

+++ Source/utils/cel_to_clx.cpp
@@ -62,7 +62,7 @@ OwnedClxSpriteListOrSheet CelToClx(const uint8_t *data, size_t size, PointerOrVa
 			numFrames = maybeNumFrames;
 		} else {
 			numFrames = LoadLE32(data);
-			WriteLE32(&cl2Data[4 * group], cl2Data.size());
+			WriteLE32(&cl2Data[4 * group], static_cast<uint32_t>(cl2Data.size()));
 		}
 
 		// CL2 header: frame count, frame offset for each frame, file size
@@ -110,7 +110,7 @@ OwnedClxSpriteListOrSheet CelToClx(const uint8_t *data, size_t size, PointerOrVa
 				}
 				++frameHeight;
 			}
-			WriteLE16(&cl2Data[frameHeaderPos + 4], frameHeight);
+			WriteLE16(&cl2Data[frameHeaderPos + 4], static_cast<uint16_t>(frameHeight));
 			memset(&cl2Data[frameHeaderPos + 6], 0, 4);
 			AppendClxTransparentRun(transparentRunWidth, cl2Data);
 		}
